# No longer run app while running test suite. To make test suite more reliable, especially on Travis