Works locally.

See how it works with Travis.

---

Strangely, on a single, anecdote run of `make test`, it's actually slightly slower! It seems faster from a IDE. But let's see if this helps make it run on Travis more reliably.

Before PR
---
Fri Feb 17 13:50:05 +08 2023
$ make test
Tests Passed: 0 failed, 349 total (314.398 seconds)
Fri Feb 17 13:56:39 +08 2023
so 6m 34s


After PR
---
Fri Feb 17 13:35:35 +08 2023
$ make test
Tests Passed: 0 failed, 349 total (327.396 seconds)
Fri Feb 17 13:42:30 +08 2023
so 6m 55s